Fix issues related to page links and CSS
Moved the 3 website pages from the subdirectory "Recipes" they were previously in in an attempt to get the style.css effects to be applied. Also corrected the respective image links for each html file to take into account this move so that the reference made to the "images" folder would be done with respect to the html recipe files NOW being in the root folder along with everything else.